# app-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HobbyHub 后端 API 文档",
        "title": "HobbyHub API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
...
        "/v1/friend": {
            "get": {
                "description": "通过用户ID获取好友列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "好友相关接口"
                ],
                "summary": "获取好友列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "通过ID更新好友状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "好友相关接口"
                ],
                "summary": "更新好友状态",
                "parameters": [
                    {
                        "description": "好友申请信息",
                        "name": "updatefriendRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateFriendRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Friend"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "通过用户ID发送好友申请",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "好友相关接口"
                ],
                "summary": "发送好友申请",
                "parameters": [
                    {
                        "description": "好友申请信息",
                        "name": "friendRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.FriendRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/friend/{id}": {
            "delete": {
                "description": "通过ID删除好友状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "好友相关接口"
                ],
                "summary": "删除好友",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "需要删除的Friend 记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.FriendRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "description": "好友用户ID",
                    "type": "integer"
                }
            }
        },
...
        "api.UpdateFriendRequest": {
            "type": "object",
            "required": [
                "friend_id",
                "status"
            ],
            "properties": {
                "friend_id": {
                    "description": "好友ID",
                    "type": "integer"
                },
                "status": {
                    "description": "申请状态，0-拒绝，1-同意",
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorMessage": {
                    "type": "string"
                }
            }
        },
...
        "models.Friend": {
            "type": "object",
            "properties": {
                "create_time": {
                    "type": "string"
                },
                "friend_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "successMessage": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "addr": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "headImg": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}